requirements.txt
crispy-bootstrap5==0.7
Django==4.2.6
django-crispy-forms==2.1
django-pandas==0.6.6
django-recaptcha==3.0.0
openpyxl==3.1.2
pandas==2.1.1
scikit-learn==1.3.2